+++ apps/restconf/restconf_lib.c
@@ -355,7 +355,6 @@ get_user_cookie(char  *cookiestr,
 int
 restconf_terminate(clixon_handle h)
 {
-    yang_stmt *yspec;
     cvec      *nsctx;
     cxobj     *x;
     int        fs; /* fgcx socket */
@@ -366,10 +365,6 @@ restconf_terminate(clixon_handle h)
     /* Delete all plugins, and RPC callbacks */
     clixon_plugin_module_exit(h);
     clicon_rpc_close_session(h);
-    if ((yspec = clicon_dbspec_yang(h)) != NULL)
-        ys_free(yspec);
-    if ((yspec = clicon_config_yang(h)) != NULL)
-        ys_free(yspec);
     yang_exit(h);
     if ((nsctx = clicon_nsctx_global_get(h)) != NULL)
         cvec_free(nsctx);