+++ lib/src/clixon_datastore_read.c
@@ -702,7 +702,7 @@ xmldb_readfile(clixon_handle    h,
              * Same ymodules are inserted into yspec1, ie pointers only
              */
             if (needclone && xmodfile){
-                if ((yspec1 = yspec_new()) == NULL)
+                if ((yspec1 = yspec_new(h, "tmp")) == NULL)
                     goto done;
                 xmsd = NULL;
                 while ((xmsd = xml_child_each(xmodfile, xmsd, CX_ELMNT)) != NULL) {